+++ example/a.ts
@@ -1,4 +1,14 @@
 import { Elysia, t } from '../src'
-import { req } from '../test/utils'
 
-const plugin = new Elysia().get('/plugin', 'static ')
+new Elysia()
+	.ws('/', {
+		open: (ws) => {
+			ws.publish('channel', 'hello')
+		},
+		response: t.String()
+	})
+	.listen(3000)
+
+// app.handle(new Request('http://localhost'))
+// 	.then((x) => x.headers.getSetCookie())
+// 	.then(console.log)